restrict denovo kmer size

+++ include/denovo_discovery/denovo_discovery.h
@@ -18,7 +18,7 @@ public:
     bool clean_assembly_graph { false };
     const uint16_t max_insertion_size;
 
-    DenovoDiscovery(const uint32_t& kmer_size, const double& read_error_rate,
+    DenovoDiscovery(const uint16_t& kmer_size, const double& read_error_rate,
         uint16_t max_insertion_size = 15, uint16_t min_covg_for_node_in_assembly_graph = 1);
 
     void find_paths_through_candidate_region(
@@ -28,7 +28,7 @@ public:
         const uint32_t& read_covg, const uint32_t& ref_length) const;
 
 private:
-    const uint32_t kmer_size;
+    const uint16_t kmer_size;
     const double read_error_rate;
 };
